// CabsBookingSPA/src/app/Shared/models/addBooking.ts
export interface AddBooking{
    email: string;
    bookingDate: Date;
    bookingTime: string;
    pickupAddress: string;
    landmark: string;
    pickupDate: Date;
    pickupTime: string;
    contactNo: string;
    status: string;
    fromPlace: string;
    toPlace: string;
    cabTypeName: string;
}